Latin grammar: clean up and update
This patch removes all superfluous files from the directory, especially the
binary files in the doc folder. They are still available in the git
repository http://github.com/daherb/GF-latin which is still the main
repository for the development of the latin grammar.
The GF files are updated to a state close to the one at the end of the GF
Summer School 2015 where they were extended to support the MUSTE grammar.