(Ara) Fix bug in ImpersCl
ImpersCl (VP -> Cl) used to check whether the VP is predicative, and
in the positive case, force the dummy subject to not be a pronoun.
This has been wrong since the time I changed prodrop rules in PredVP.

The error that followed manifested when the Cl was put in subordinate
position: the subject pronoun did not attach to the conjunction.
An example follows.

Lang: SubjS that_Subj (UseCl (TTAnt TPres ASimul) PPos (ImpersCl (UseComp (CompNP (DetCN (DetQuant IndefArt NumSg) (UseN girl_N))))))
LangEng: that it is a girl

Old LangAra: أنَّ ها بِنتٌ (wrong)
New LangAra: أنَّ &+ ها بِنتٌ (correct)