(Pes) Remove redundant parameters in V*, VP and Cl
Keep VerbForm at VP level, and rename VPHTense to use Tense and 
Anteriority instead of custom names that are then mapped to Tense and 
Anteriority in Use*Cl functions.

WIP: Remove Aux type and use beVerb instead of beAux. There were a lot 
of comments along the lines "check the right form"; will do that with a 
native speaker.