(Eng) Add lin for QuestVPS and SQuestVPS + new lincats for VPS*
Now VPS and VPS2 need to potentially become QS, so we keep them
discontinuous. Works great with lists too, e.g.

  {fin="wouldn't" ; inf="have seen and waited for (obj)"}

NB. copula disappears, if it's a part of a diverse list of questions:

  Statements:
    you sleep or are small -- ok
    you are small or sleep -- ok

  Questions:
    ?are you small or sleep -- for "are you small or do you sleep", need to implement [QS]
    *do you sleep or small -- infinitive "be small" not retrievable from VPS currently

In a list of only predicates, this addition is rather redundant.
You can already say "are you big or small" with AP coordination.
But no other RGL construction allows "do you sleep or drink beer".