Updating num200 & num2000
The letter ت in Arabic has two cases, opened case ت and closed case ة. The closed case only appear as the last letter in a word. Since hundred is written with closed case "مئة" but two hundred require adding additional letter at the end, then this letter can't be remain in its closed case, but need to be changed to open case.
For the two thousand I just added the missing tone, it can be ignored, but since the Arabic RGL all written with tones, thus I added this one.